requirements.txt
python-telegram-bot==20.*
requests
beautifulsoup4
fake-useragent
schedule
python-dotenv
selenium
webdriver-manager